Popravljen unos rezultata

+++ Faza 5/portal za sahovski savez/database/migrations/2020_05_14_120043_create_results_table.php
@@ -18,8 +18,8 @@ class CreateResultsTable extends Migration
             $table->foreignId('tournament_id')->constrained('tournaments')->onDelete('cascade')->onUpdate('cascade');
             $table->foreignId('white_id')->constrained('players')->onDelete('no action')->onUpdate('no action');
             $table->foreignId('black_id')->constrained('players')->onDelete('no action')->onUpdate('no action');
-            $table->integer('white_result');
-            $table->integer('black_result');
+            $table->float('white_result');
+            $table->float('black_result');
             $table->integer('table');
             $table->integer('round');
             $table->foreignId('arbiter_id')->constrained('players')->onDelete('no action')->onUpdate('no action');
